Properly handle slope rescaling
The previous implementation had a bug, because although it cast each
slice's data to a float32, it then assigned it to the original array
(which would cast it back to the original data type).

This implementation assumes that all slices will have the same slope and
intercept rescaling.  It is not clear whether this is the case.  If not,
we may need to adjust the implementation in the future.