Only cast to float if rescaled
Some DICOM datasets have a different `RescaleSlope` and
`RescaleIntercept` value for each slice, thus we need to apply the
rescaling separately to each one.

If none of the slices require rescaling, preserve the initial data type;
this is because, in general, we may not want to waste time (or loose
precision) casting when we don't need to.  By allowing users to do this
upstream, we give them the option to handle it however they want.