Reverse slices to orient patient head at slice 0
The top of the patient is considered the first slice in a volume, while
the bottom of the patient is considered the last. In order to reflect
this in the slice ordering, sorting by ImagePatientPosition must be
reversed (starting from the highest point in the slice direction and
ending at the lowest).