Move version to separate module to prevent buildtime issues
Because setup.py was importing the package in order to get the version
parameter, it was hitting errors about missing build-time dependencies
before it actually got to the step of installing the build-time
dependencies.

This was resolved by separating the version out into its own module,
which is read in directly by setup.py.

See the following SO post for more information:

https://stackoverflow.com/questions/2058802/how-can-i-get-the-version-defined-in-setup-py-setuptools-in-my-package