Specify sort key when sorting by slice position
The FileDataset object in PyDicom does not support comparison
operations, so it throws an error if used in the sort function. This was
never the intended behavior anyways; we only want to sort by slice
position, and it does not matter the sort order of two slices at the
same position as long as the order is consistent. Therefore, a key
function was specified that only checks the slice position portion of
the tuple.