Allow bypassing the sort logic for custom slice sorting
For some image series types, such as diffusion-weighted MRI, multiple
scans over the same body region may be performed within a single series.
Depending on the technique and scanner, the instance numbers may not be
enough to separate out the slices from the different scans over the
body. In this case, an option has been added to `combine_slices` that
allows the user to bypass the sorting logic and perform their own
sorting prior to passing the slices in.